[ADD] git_trunk: submodule support
Introducing git trunk commands support on submodules. It is now possible
to run all commands from submodules just like on main repo.

This makes it easier to work with submodules directly from superproject.

Meaning that user won't need to switch context/projects, separately
develop dependency repository, finish it and then pull changes via
superproject submodule.

[FIX] git_trunk_commands: active_ref_name

Using it be able to always get reference, when needed (when repo might
not be on active branch, but on detached head like tag or raw commit).

[REF] .travis: use unittest over nose

nose package is obsolete now.

[REF] git_trunk: somewhat better typing coverage

[REF] git_trunk: split git_trunk.py

Splitting git_trunk.py for better readability, separating concerns,
reusability.
