+++ .gitignore
@@ -7,3 +7,5 @@
 build/
 dist/
 MANIFEST
+# mypy
+.mypy_cache/
