+++ .travis.yml
@@ -14,4 +14,5 @@ install:
   - "pip install ."
 
 # command to run tests
-script: nosetests
+script:
+ - python -m unittest discover -v
