+++ git_trunk/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '0.3.0'
+__version__ = '0.4.0'
